pyproject.toml
[project]
name = "plaud-mcp"
version = "0.1.0"
description = "MCP server for Plaud transcripts via direct API with macOS Keychain auth"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.25.0",
    "httpx>=0.27.0",
    "cryptography>=43.0",
]

[project.scripts]
plaud-mcp = "plaud_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.400",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.5",
]

[project.optional-dependencies]
audio = [
    "mlx-whisper>=0.4.0",
    "pyannote.audio>=3.1.0",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
